+++ R/sidebar.R
@@ -20,49 +20,52 @@ sidebar_deps <- function(mini = FALSE) {
 #' @param map A leaflet map widget
 #' @param id Id of the sidebar-div. Must match with the \code{id} of
 #'   \code{\link{sidebar_tabs}}
-#' @param options A named list with \code{position} and \code{fit} elements.
+#' @param options A named list with the only option \code{position}, which should
+#'   be either \code{left} or \code{right}.
 #' @family Sidebar Functions
 #' @references \url{https://github.com/Turbo87/sidebar-v2}
 #' @export
 #' @inherit leaflet::addControl return
-#' @examples \dontrun{
-#' library(shiny)
-#' runApp(paste0(system.file("examples", package = "leaflet.extras2"),
-#'               "/sidebar_app.R"))
-#' }
+#' @inherit sidebar_tabs examples
 addSidebar <- function(map, id = "sidebar",
-                       options = list(position = "left",
-                                      fit = TRUE)) {
+                       options = list(position = "left")) {
   map$dependencies <- c(map$dependencies, sidebar_deps())
+  options$fit = TRUE
   invokeMethod(map, NULL, "addSidebar", id, options)
 }
 
 #' Remove the Sidebar
 #' @param map A leaflet map widget
+#' @param sidebar_id The id of the sidebar (per \code{\link{sidebar_tabs}}).
+#'  Defaults to \code{NULL} such that the first sidebar is opened.
 #' @family Sidebar Functions
 #' @inherit leaflet::addControl return
 #' @export
-removeSidebar <- function(map) {
-  invokeMethod(map, NULL, "removeSidebar")
+removeSidebar <- function(map, sidebar_id = NULL) {
+  invokeMethod(map, NULL, "removeSidebar", sidebar_id)
 }
 
 #' Close the Sidebar
 #' @param map A leaflet map widget
+#' @param sidebar_id The id of the sidebar (per \code{\link{sidebar_tabs}}).
+#'  Defaults to \code{NULL} such that the first sidebar is opened.
 #' @family Sidebar Functions
 #' @inherit leaflet::addControl return
 #' @export
-closeSidebar <- function(map) {
-  invokeMethod(map, NULL, "closeSidebar")
+closeSidebar <- function(map, sidebar_id = NULL) {
+  invokeMethod(map, NULL, "closeSidebar", sidebar_id)
 }
 
 #' Open the Sidebar by ID
 #' @param map A leaflet map widget
-#' @param id The id of the \code{\link{sidebar_pane}} to open
+#' @param id The id of the \code{\link{sidebar_pane}} to open.
+#' @param sidebar_id The id of the sidebar (per \code{\link{sidebar_tabs}}).
+#'  Defaults to \code{NULL} such that the first sidebar is opened.
 #' @family Sidebar Functions
 #' @inherit leaflet::addControl return
 #' @export
-openSidebar <- function(map, id) {
-  invokeMethod(map, NULL, "openSidebar", id)
+openSidebar <- function(map, id, sidebar_id = NULL) {
+  invokeMethod(map, NULL, "openSidebar", list(id = id, sidebar_id = sidebar_id))
 }
 
 #' Create a Sidebar Pane
@@ -103,8 +106,14 @@ sidebar_pane <- function(title = "Sidebar Title",
 #' @export
 #' @examples \dontrun{
 #' library(shiny)
+#'
+#' # run example app showing a single sidebar
 #' runApp(paste0(system.file("examples", package = "leaflet.extras2"),
 #'               "/sidebar_app.R"))
+#'
+#' # run example app showing two sidebars
+#' runApp(paste0(system.file("examples", package = "leaflet.extras2"),
+#'               "/multi_sidebar_app.R"))
 #' }
 sidebar_tabs <- function(id = "sidebar", iconList = NULL, ...){
   arg <- list(...)
@@ -112,14 +121,14 @@ sidebar_tabs <- function(id = "sidebar", iconList = NULL, ...){
   if (length(ids) != length(iconList))
     stop("The number of icons needs to match the number of sidebar panes.")
   tags$div(id = id, class = "sidebar collapsed",
-           tags$div(class = "sidebar-tabs",
+           tags$div(class = "sidebar-tabs", style = "display: none",
                     tags$ul(role = "tablist",
                             tagList(lapply(1:length(ids), function(x) {
                               tags$li(tags$a(href = paste0("#", ids[[x]]), role = "tab", iconList[[x]]))
                             }))
                             )
            ),
-           tags$div(class = "sidebar-content",
+           tags$div(class = "sidebar-content", style = "display: none",
                     tagList(arg))
   )
 }